Benchmarks

+++ src/SourceCode.Clay/BitOps.Core.cs
@@ -65,7 +65,7 @@ namespace SourceCode.Clay
             return Unsafe.AddByteOffset(
                 ref MemoryMarshal.GetReference(s_TrailingZeroCountDeBruijn),
                 // Using deBruijn sequence, k=2, n=5 (2^5=32) : 0b_0000_0111_0111_1100_1011_0101_0011_0001u
-                (IntPtr)(((uint)((value & -value) * 0x077CB531u)) >> 27));
+                (IntPtr)(((value & -value) * 0x077CB531u) >> 27));
         }
 
         /// <summary>
